Refactor again to reduce yet another circular dependency coming from interpolation routines. Also add MPI to unit tests so that they actually run when compiled with MPI.
